+++ app/src/main/java/com/example/sport_planet/data/response/EvaluateListResponse.kt
@@ -1,4 +1,4 @@
-package com.example.sport_planet.model
+package com.example.sport_planet.data.response
 
 data class EvaluateListResponse(
     val status : Int,
