+++ app/src/main/java/com/example/sport_planet/data/response/MyViewEditRequest.kt
@@ -1,7 +1,5 @@
-package com.example.sport_planet.model
+package com.example.sport_planet.data.response
 
-
-import com.google.gson.annotations.SerializedName
 
 data class MyViewEditRequest(
     val address: String,
