+++ app/src/main/java/com/example/sport_planet/data/response/MyViewHistoryResponse.kt
@@ -1,6 +1,7 @@
-package com.example.sport_planet.model
+package com.example.sport_planet.data.response
 
 
+import com.example.sport_planet.data.model.MyViewHistoryModel
 import com.google.gson.annotations.SerializedName
 
 data class MyViewHistoryResponse(
