+++ app/src/main/java/com/example/sport_planet/data/response/TagResponse.kt
@@ -1,0 +1,5 @@
+package com.example.sport_planet.data.response
+
+data class TagResponse(
+    val id: Long
+)
