--- app/src/main/java/com/example/sport_planet/model/TagResponse.kt
@@ -1,5 +1,0 @@
-package com.example.sport_planet.model
-
-data class TagResponse(
-    val id: Long
-)
