+++ app/src/main/java/com/example/sport_planet/remote/RemoteDataSource.kt
@@ -1,8 +1,7 @@
 package com.example.sport_planet.remote
 
-import com.example.sport_planet.model.*
+import com.example.sport_planet.data.response.*
 import io.reactivex.Single
-import retrofit2.http.Body
 
 interface RemoteDataSource{
     fun getExercise() : Single<ExerciseResponse>
